chore(inv-compliance): move files to suit the v3 build
