+++ packages/inventory-compliance/src/Compliance/index.js
@@ -1,0 +1,2 @@
+export { default } from './Compliance';
+export * from './Compliance';
