+++ packages/inventory-compliance/src/ComplianceEmptyState/index.js
@@ -1,0 +1,2 @@
+export { default } from './ComplianceEmptyState';
+export * from './ComplianceEmptyState';
