+++ packages/inventory-compliance/src/ComplianceRemediationButton/index.js
@@ -1,0 +1,2 @@
+export { default } from './ComplianceRemediationButton';
+export * from './ComplianceRemediationButton';
