+++ packages/inventory-compliance/src/Constants/index.js
@@ -1,0 +1,2 @@
+export { default } from './Constants';
+export * from './Constants';
