+++ packages/inventory-compliance/src/Fixtures/Fixtures.js
