+++ packages/inventory-compliance/src/Fixtures/index.js
@@ -1,0 +1,2 @@
+export { default } from './Fixtures';
+export * from './Fixtures';
