+++ packages/inventory-compliance/src/SystemPolicyCard/index.js
@@ -1,0 +1,2 @@
+export { default } from './SystemPolicyCard';
+export * from './SystemPolicyCard';
