+++ packages/inventory-compliance/src/SystemPolicyCards/index.js
@@ -1,0 +1,2 @@
+export { default } from './SystemPolicyCards';
+export * from './SystemPolicyCards';
