+++ packages/inventory-compliance/src/SystemRulesTable/index.js
@@ -1,0 +1,2 @@
+export { default } from './SystemRulesTable';
+export * from './SystemRulesTable';
