+++ packages/inventory-compliance/src/index.scss
@@ -1,0 +1,1 @@
+@import "./Compliance//compliance.scss";
