+++ packages/inventory/package.json
@@ -41,7 +41,7 @@
     "dependencies": {
         "@redhat-cloud-services/frontend-components": "^2.0.0",
         "@redhat-cloud-services/frontend-components-notifications": "^2.0.0",
-        "@redhat-cloud-services/frontend-components-utilities": "*",
+        "@redhat-cloud-services/frontend-components-utilities": "^2.0.0",
         "@redhat-cloud-services/host-inventory-client": "1.0.91",
         "axios": "^0.21.1"
     }
